Add finance_calculator.py

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,7 @@
 # Prompt user to insert their monthly income
-monthly_income = int(input("Enter your monthly income: "))
+monthly_income = float(input("Enter your monthly income: "))
 # Prompt user to insert their total monthly expenses
-total_monthly_expenses = int(input("Enter your total monthly expenses: "))
+total_monthly_expenses = float(input("Enter your total monthly expenses: "))
 # Calculate monthly savings
 monthly_savings = monthly_income - total_monthly_expenses
 # Print the result
